Dashboard Refresh & Code Reorganization
A brand new dashboard with user authentication, registration via email confirmation, oAuth, password recovery, and captcha protection, and Soft Deletes for users

+++ database/migrations/2017_10_30_223026_bible_translators_table.php
@@ -13,35 +13,42 @@ class BibleTranslatorsTable extends Migration
 	 */
 	public function up()
 	{
-		Schema::create('translators', function (Blueprint $table) {
-			$table->string('id',191)->primary()->unique();
-			$table->string('name');
-			$table->string('born')->nullable();
-			$table->string('died')->nullable();
-			$table->text('description');
-			$table->timestamps();
-		});
 
-		Schema::create('bible_translator', function($table) {
-			$table->string('bible_id',12)->index();
-			$table->foreign('bible_id')->references('id')->on('bibles')->onUpdate('cascade')->onDelete('cascade');
-			$table->string('translator_id',191);
-			$table->foreign('translator_id')->references('id')->on('translators')->onDelete('cascade')->onUpdate('cascade');
-			$table->timestamps();
-		});
+		if(!Schema::connection('dbp')->hasTable('translators')) {
+			Schema::connection('dbp')->create('translators', function (Blueprint $table) {
+				$table->string('id', 191)->primary()->unique();
+				$table->string('name');
+				$table->string('born')->nullable();
+				$table->string('died')->nullable();
+				$table->text('description');
+				$table->timestamps();
+			});
+		}
 
-		Schema::create('translator_relations', function (Blueprint $table) {
-			$table->string('translator_id',191);
-			$table->foreign('translator_id')->references('id')->on('translators')->onDelete('cascade')->onUpdate('cascade');
-			$table->string('translator_relation_id',191);
-			$table->foreign('translator_relation_id')->references('id')->on('translators')->onUpdate('cascade')->onDelete('cascade');
-			$table->integer('organization_id')->unsigned()->nullable();
-			$table->foreign('organization_id')->references('id')->on('organizations');
-			$table->string('type');
-			$table->string('description');
-			$table->string('notes');
-			$table->timestamps();
-		});
+		if(!Schema::connection('dbp')->hasTable('bible_translator')) {
+			Schema::connection('dbp')->create('bible_translator', function ($table) {
+				$table->string('bible_id', 12)->index();
+				$table->foreign('bible_id')->references('id')->on('bibles')->onUpdate('cascade')->onDelete('cascade');
+				$table->string('translator_id', 191);
+				$table->foreign('translator_id')->references('id')->on('translators')->onDelete('cascade')->onUpdate('cascade');
+				$table->timestamps();
+			});
+		}
+
+		if(!Schema::connection('dbp')->hasTable('translator_relations')) {
+			Schema::connection('dbp')->create('translator_relations', function (Blueprint $table) {
+				$table->string('translator_id', 191);
+				$table->foreign('translator_id')->references('id')->on('translators')->onDelete('cascade')->onUpdate('cascade');
+				$table->string('translator_relation_id', 191);
+				$table->foreign('translator_relation_id')->references('id')->on('translators')->onUpdate('cascade')->onDelete('cascade');
+				$table->integer('organization_id')->unsigned()->nullable();
+				$table->foreign('organization_id')->references('id')->on('organizations');
+				$table->string('type');
+				$table->string('description');
+				$table->string('notes');
+				$table->timestamps();
+			});
+		}
 
 	}
 
@@ -52,8 +59,8 @@ class BibleTranslatorsTable extends Migration
 	 */
 	public function down()
 	{
-		Schema::dropIfExists('translator_relations');
-		Schema::dropIfExists('bible_translator');
-		Schema::dropIfExists('translators');
+		Schema::connection('dbp')->dropIfExists('translator_relations');
+		Schema::connection('dbp')->dropIfExists('bible_translator');
+		Schema::connection('dbp')->dropIfExists('translators');
 	}
 }
